requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.3
pydantic-settings==2.8.1
sqlalchemy==2.0.40
textual==1.0.0
rich==14.0.0
httpx==0.28.1
python-multipart==0.0.20
PyYAML==6.0.2
pytest==8.3.5
